feat:✨ renew list_installed_mod function
Changes
- now checksum of `LocalModinfo` become `Option<String>`
- removed hashing task from `list_installed_mods`.
    now it takes only 200ms for 137 mods (about 3.1GB) before it took over 700ms.
- move file realted functions from src/download.rs to src/fileutil.rs
- move local mods related functions from src/download.rs to src/installed_mod.rs
- add new error: FileIsNotHashing
- add `update_mod_hashes` function to set checksums to installed mods
